Refs #28900 - in MainLerntiaControler initialisation, PROPERTIES are checked
Method checkIfAllPropertiesFilesAreProvided implemented in ConfigReader
class. This method is called when MainLerntiaControler is initalised and
it warnes the user that not all properties files are provided;

Implemented exception handeling in ConfigReader via a newly implemented
ConfigReaderException;

Changed all java classes and junit tests so that they can handle the
newly created ConfigReaderException;

Updated JavaDoc in service interfaces, in order to proclame that some
methods now also throw ServiceException if properties file is not provided;

Added additional LOG debug messages to enable a bettter tracking of
PROPERTIES checking.